Extract parser into its own gem.
