+++ jsonapi_parser/lib/jsonapi/parse.rb
@@ -1,5 +1,19 @@
 require 'json'
 
+require 'jsonapi/attributes'
+require 'jsonapi/document'
+require 'jsonapi/error'
+require 'jsonapi/exceptions'
+require 'jsonapi/jsonapi'
+require 'jsonapi/link'
+require 'jsonapi/links'
+require 'jsonapi/relationship'
+require 'jsonapi/relationships'
+require 'jsonapi/resource'
+require 'jsonapi/resource_identifier'
+require 'jsonapi/parser/document'
+require 'jsonapi/parser/exceptions'
+
 module JSONAPI
   module_function
 
@@ -13,6 +27,6 @@ module JSONAPI
   def parse(document, options = {})
     hash = document.is_a?(Hash) ? document : JSON.parse(document)
 
-    Document.new(hash, options)
+    Parser::Document.new(hash, options)
   end
 end
