+++ jsonapi_parser/lib/jsonapi/parser/exceptions.rb
@@ -1,0 +1,6 @@
+module JSONAPI
+  module Parser
+    class InvalidDocument < StandardError
+    end
+  end
+end
