+++ jsonapi_parser/spec/duplicates_spec.rb
@@ -1,6 +1,6 @@
-require 'jsonapi'
+require 'jsonapi/parser'
 
-describe JSONAPI, '#parse' do
+describe JSONAPI::Parser, '#parse' do
   it 'succeeds when there are no duplicates' do
     payload = {
       'data' => [
@@ -38,7 +38,7 @@ describe JSONAPI, '#parse' do
         }]
     }
 
-    JSONAPI.parse(payload, verify_duplicates: true)
+    JSONAPI::Parser.parse(payload, verify_duplicates: true)
   end
 
   it 'fails when there are duplicates within primary data' do
@@ -53,8 +53,8 @@ describe JSONAPI, '#parse' do
         }]
     }
 
-    expect { JSONAPI.parse(payload, verify_duplicates: true) }
-      .to raise_error(JSONAPI::InvalidDocument)
+    expect { JSONAPI::Parser.parse(payload, verify_duplicates: true) }
+      .to raise_error(JSONAPI::Parser::InvalidDocument)
   end
 
   it 'fails when there are duplicates within included' do
@@ -70,8 +70,8 @@ describe JSONAPI, '#parse' do
         }]
     }
 
-    expect { JSONAPI.parse(payload, verify_duplicates: true) }
-      .to raise_error(JSONAPI::InvalidDocument)
+    expect { JSONAPI::Parser.parse(payload, verify_duplicates: true) }
+      .to raise_error(JSONAPI::Parser::InvalidDocument)
   end
 
   it 'fails when there are duplicates within primary data' do
@@ -89,7 +89,7 @@ describe JSONAPI, '#parse' do
         }]
     }
 
-    expect { JSONAPI.parse(payload, verify_duplicates: true) }
-      .to raise_error(JSONAPI::InvalidDocument)
+    expect { JSONAPI::Parser.parse(payload, verify_duplicates: true) }
+      .to raise_error(JSONAPI::Parser::InvalidDocument)
   end
 end
