+++ jsonapi_parser/spec/parser_spec.rb
@@ -1,6 +1,6 @@
-require 'jsonapi'
+require 'jsonapi/parser'
 
-describe JSONAPI, '#parse' do
+describe JSONAPI::Parser, '#parse' do
   before(:all) do
     @author_links_hash = {
       'self' => 'http://example.com/articles/1/relationships/author',
@@ -47,7 +47,7 @@ describe JSONAPI, '#parse' do
   end
 
   it 'works' do
-    document = JSONAPI.parse(@payload)
+    document = JSONAPI::Parser.parse(@payload)
 
     expect(document.meta.keys).to eq ['count']
     expect(document.meta['count']).to eq '13'
