--- lib/jsonapi/exceptions.rb
@@ -1,4 +1,0 @@
-module JSONAPI
-  class InvalidDocument < StandardError
-  end
-end
