--- lib/jsonapi/version.rb
@@ -1,3 +1,0 @@
-module JSONAPI
-  VERSION = '0.1.1.beta2'.freeze
-end
